pyproject.toml
[tool.poetry]
name = "ai-assistant"
version = "0.1.0"
description = "backend server to interact with ai models"
authors = ["dpills"]
readme = "README.md"

[tool.poetry.scripts]
start = "main:app"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.2.3"
python-dotenv = "^1.0.0"
fastapi = "^0.104.1"
uvicorn = { extras = ["standard"], version = "^0.24.0.post1" }
websockets = "^12.0"
databases = "^0.9.0"
asyncpg = "^0.29.0"
sentence-transformers = "^3.2.0"
psycopg2 = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = [
    "I",
    "RUF",
    "W",
    "C90",
    "UP"
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"